# cta-bus-tracker-openapi.yaml
openapi: 3.0.3
info:
  title: CTA Bus Tracker API
  description: >
    The Bus Tracker API allows for querying data from the CTA Bus Tracker service.
    The system can provide estimated arrival times for buses as they approach bus stops,
    as well as data on routes, stops, patterns, detours, cancelations and more.
    Information about buses and their estimated arrivals is updated about once every 30 seconds.

    **Authentication:** All requests require a valid 25-digit API key passed via the `key` query parameter.

    **Response Format:** XML is the default. Pass `format=json` to receive JSON responses.
    This spec documents the JSON format only.

    **Rate Limiting:** By default, one API key can make a maximum of 100,000 requests per day.

    **Versioning:** This document covers Version 3 only. The version is embedded in the URL path as `/v3/`.

    **Legal:** By using this API, you agree to the CTA Developer Terms of Use.
    This service is provided on an as-is basis without guarantees as to availability or accuracy.
  version: "3.0"
  contact:
    name: Chicago Transit Authority
    url: https://www.transitchicago.com
  license:
    name: CTA Developer Terms of Use
    url: https://www.transitchicago.com/developers/terms
servers:
  - url: https://www.ctabustracker.com/bustime/api/v3
    description: CTA Bus Tracker API v3 (port 80)
...
paths:
...
  /getroutes:
    get:
      operationId: getRoutes
      summary: Get all routes
      description: >
        Retrieve the set of routes serviced by the system.
        Returns route designators, names, colors, and display designators.
      tags:
        - Routes
      parameters:
        - $ref: '#/components/parameters/key'
        - $ref: '#/components/parameters/format'
        - $ref: '#/components/parameters/locale'
      responses:
        '200':
          description: Successful response containing all available routes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRoutesResponse'
              examples:
                routes:
                  summary: Sample route list
                  value:
                    bustime-response:
                      routes:
                        - rt: "1"
                          rtnm: "Pontiac – Dhu Varren"
                          rtdd: "1"
                          rtclr: "#ffffff"
                        - rt: "2"
                          rtnm: "Pontiac - University"
                          rtdd: "1"
                          rtclr: "#dc78af"
        '400':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────────
  # GET DIRECTIONS
  # ─────────────────────────────────────────────
  /getdirections:
    get:
      operationId: getDirections
      summary: Get directions for a route
      description: >
        Retrieve the set of directions serviced by the specified route.
        The returned `id` field is what must be passed as the `dir` parameter in
        subsequent calls such as `getstops`.

        **Note:** In v3, the Route Directions call always returns locale-specific data
        regardless of whether the system is single-feed or multi-feed.
      tags:
        - Routes
      parameters:
        - $ref: '#/components/parameters/key'
        - $ref: '#/components/parameters/format'
        - $ref: '#/components/parameters/locale'
        - name: rt
          in: query
          required: true
          description: Alphanumeric designator of a route (e.g. `"20"` or `"X20"`) for which directions are to be returned.
          schema:
            type: string
            example: "20"
      responses:
        '200':
          description: Successful response containing the available directions for the specified route.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDirectionsResponse'
              examples:
                route20Directions:
                  summary: Directions for Route 20
                  value:
                    bustime-response:
                      directions:
                        - id: "Eastbound"
                          name: "Eastbound"
                        - id: "Westbound"
                          name: "Westbound"
        '400':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────────
  # GET STOPS
  # ─────────────────────────────────────────────
  /getstops:
    get:
      operationId: getStops
      summary: Get stops for a route/direction or by stop IDs
      description: >
        Retrieve the set of stops for the specified route and direction,
        or retrieve one or more specific stops by stop ID.

        A request must provide either a `rt` + `dir` pair OR up to 10 `stpid`s — not both.

        Stop lists are only available for a valid route/direction pair.
        A list of all stops serving a particular route regardless of direction cannot be requested.

        If a stop is affected by a detour, the detour's ID will appear in:
        - `dtradd` — if the stop was temporarily added to the pattern
        - `dtrrem` — if the stop has been detoured around (removed)

        Use `getdetours` to retrieve metadata about a referenced detour ID.
      tags:
        - Stops
      parameters:
        - $ref: '#/components/parameters/key'
        - $ref: '#/components/parameters/format'
        - $ref: '#/components/parameters/locale'
        - name: rt
          in: query
          required: false
          description: >
            Alphanumeric designator of the route (e.g. `"20"` or `"X20"`) for which stops are to be returned.
            Required if `stpid` is not provided.
          schema:
            type: string
            example: "20"
        - name: dir
          in: query
          required: false
          description: >
            Direction of the route (e.g. `"Eastbound"`). Must match the direction `id`
            from the `getdirections` call. Required if `stpid` is not provided.
          schema:
            type: string
            example: "Eastbound"
        - name: stpid
          in: query
          required: false
          description: >
            Comma-delimited list of numeric stop IDs (e.g. `"305,1577,1564"`).
            Up to 10 stop IDs can be specified. Required if `rt` and `dir` are not provided.
          schema:
            type: string
            example: "1577,1564"
      responses:
        '200':
          description: Successful response containing stop data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStopsResponse'
              examples:
                stopsByRouteDir:
                  summary: Stops for Route 20 Eastbound
                  value:
                    bustime-response:
                      stops:
                        - stpid: "1577"
                          stpnm: "1509 S Michigan"
                          lat: 41.861706666665
                          lon: -87.623969999999
                        - stpid: "1564"
                          stpnm: "3000 S Michigan"
                          lat: 41.840606666667
                          lon: -87.623206666667
                          dtrrem:
                            - "BFC46F62-990F-4AB4-A85C-3AF84574EC99"
                            - "50C633C7-0891-4E5A-83A8-FF0C6214BF69"
        '400':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────────
  # GET PATTERNS
  # ─────────────────────────────────────────────
...
  /getpredictions:
    get:
      operationId: getPredictions
      summary: Get arrival/departure predictions for stops or vehicles
      description: >
        Retrieve arrival or departure predictions for one or more stops or one or more vehicles.
        Predictions are always returned in ascending order by `prdtm`.

        Use `stpid` to retrieve predictions for one or more stops (max 10).
        Use `vid` to retrieve predictions for one or more vehicles (max 10).

        **Note:** `stpid` and `vid` cannot be combined. If both are specified, only the first
        parameter encountered will be processed. Calls without either parameter are not allowed.

        The optional `rt` parameter can filter predictions by route(s) and is only available
        when using `stpid`.

        Use `top` to limit the number of predictions returned.

        Predictions for canceled stops will not be returned unless they are configured
        to be displayed publicly. Same applies to expressed stops.

        The `dyn` field indicates if a prediction has been affected by Disruption Management.
        Refer to the Dynamic Action Types section for interpretation.
      tags:
        - Predictions
      parameters:
        - $ref: '#/components/parameters/key'
        - $ref: '#/components/parameters/format'
        - $ref: '#/components/parameters/locale'
        - name: stpid
          in: query
          required: false
          description: >
            Comma-delimited list of stop IDs (e.g. `"5029,1392,2019,4367"`).
            Maximum of 10 identifiers. Cannot be combined with `vid`.
          schema:
            type: string
            example: "456"
        - name: vid
          in: query
          required: false
          description: >
            Comma-delimited list of vehicle IDs (e.g. `"509,392,201,4367"`).
            Maximum of 10 identifiers. Cannot be combined with `stpid`.
          schema:
            type: string
            example: "8184,8140"
        - name: rt
          in: query
          required: false
          description: >
            Comma-delimited list of route designators to filter predictions.
            Only available when used with `stpid`.
          schema:
            type: string
            example: "20"
        - name: top
          in: query
          required: false
          description: Maximum number of predictions to return.
          schema:
            type: integer
            example: 5
        - name: tmres
          in: query
          required: false
          description: >
            Resolution of timestamps in the response.
            - `s` — resolution to the second (format: `YYYYMMDD HH:MM:SS`)
            - `m` — resolution to the minute (format: `YYYYMMDD HH:MM`, default)
          schema:
            type: string
            enum: [s, m]
            default: m
        - name: unixTime
          in: query
          required: false
          description: >
            If true, `tmstmp` and `prdtm` timestamps are returned as Unix epoch
            milliseconds instead of formatted date strings.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response containing prediction data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPredictionsResponse'
              examples:
                predictionsForStop:
                  summary: Predictions for stop 456 on route 20
                  value:
                    bustime-response:
                      prd:
                        - tmstmp: "20250421 16:04"
                          typ: "A"
                          stpnm: "Madison & Jefferson"
                          stpid: "456"
                          vid: "8184"
                          dstp: 686
                          rt: "20"
                          rtdd: "20"
                          rtdir: "Westbound"
                          des: "Austin"
                          prdtm: "20250421 16:06"
                          tablockid: "20 -803"
                          tatripid: "1040713"
                          origtatripno: "262522629"
                          dly: false
                          dyn: 0
                          prdctdn: "DUE"
                          zone: ""
                          psgld: ""
                          stst: 57120
                          stsd: "2025-04-21"
                          flagstop: 0
        '400':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────────
  # GET LOCALES
  # ─────────────────────────────────────────────
...
components:
...
  parameters:
    key:
      name: key
      in: query
      required: true
      description: 25-digit Bus Tracker API access key assigned to the developer.
      schema:
        type: string
        example: "89dj2he89d8j3j3ksjhdue93j"
    format:
      name: format
      in: query
      required: false
      description: >
        Response format. XML is the default if not specified.
        Pass `json` for JSON responses (documented in this spec).
      schema:
        type: string
        enum: [xml, json]
        default: xml
        example: "json"
    locale:
      name: locale
      in: query
      required: false
      description: >
        The language that the response should be in. Must be a valid ISO 639 language code
        (e.g. `"en"` for English, `"es"` for Spanish). Use `getlocalelist` to retrieve
        supported values. If the specified locale is not supported, the TA's default language
        is used without indication.
      schema:
        type: string
        example: "en"

  # ─────────────────────────────────────────────
  # RESPONSES (shared)
  # ─────────────────────────────────────────────
  responses:
    ErrorResponse:
      description: >
        Error response. Errors are returned inside the `bustime-response` object
        with an `error` array containing one or more error messages.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noServiceScheduled:
              summary: No service scheduled error
              value:
                bustime-response:
                  error:
                    - msg: "No service scheduled"
            invalidKey:
              summary: Invalid API key error
              value:
                bustime-response:
                  error:
                    - msg: "Invalid API access key supplied"
            transactionLimit:
              summary: Daily transaction limit exceeded
              value:
                bustime-response:
                  error:
                    - msg: "Transaction limit for current day has been exceeded."
            noParameter:
              summary: Missing required parameter
              value:
                bustime-response:
                  error:
                    - msg: "No parameter provided"

  # ─────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      description: An error message returned when request processing fails.
      required:
        - msg
      properties:
        msg:
          type: string
          description: Human-readable error message.
          example: "No service scheduled"
        rt:
          type: string
          description: The route designator related to the error, if applicable.
          example: "20"
        rtdir:
          type: string
          description: The route direction related to the error, if applicable.
          example: "Eastbound"
        stpid:
          type: string
          description: The stop ID related to the error, if applicable.
          example: "456"
        vid:
          type: string
          description: The vehicle ID related to the error, if applicable.
          example: "509"
        rtpidatafeed:
          type: string
          description: The RTPI data feed identifier related to the error, if applicable.
    ErrorResponse:
      type: object
      properties:
        bustime-response:
          type: object
          properties:
            error:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'

    # ── GET TIME ──
...
    Route:
      type: object
      description: A single route serviced by the system.
      required:
        - rt
        - rtnm
        - rtclr
        - rtdd
      properties:
        rt:
          type: string
          description: Alphanumeric route designator (e.g. `"20"` or `"X20"`).
          example: "1"
        rtnm:
          type: string
          description: Common name of the route (e.g. `"Madison"` for route 20).
          example: "Indiana/Hyde Park"
        rtclr:
          type: string
          description: Color of the route line used in map displays, as a hex color code.
          example: "#dc78af"
        rtdd:
          type: string
          description: Language-specific route designator intended for display.
          example: "1"
        rtpidatafeed:
          type: string
          description: RTPI data feed identifier (multi-feed systems only).
    GetRoutesResponse:
      type: object
      description: Root response wrapper for getroutes.
      properties:
        bustime-response:
          type: object
          properties:
            routes:
              type: array
              description: >
                Array of route objects. Note: in JSON the key is `routes` (plural),
                unlike the XML element name `route`.
              items:
                $ref: '#/components/schemas/Route'
            error:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'

    # ── DIRECTION ──
    Direction:
      type: object
      description: A direction of travel available for a specific route.
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: >
            Direction designator to be used in other API requests such as `getstops`.
            This is the value to pass as the `dir` parameter.
          example: "Eastbound"
        name:
          type: string
          description: Human-readable, locale-dependent name of the direction.
          example: "Eastbound"
    GetDirectionsResponse:
      type: object
      description: Root response wrapper for getdirections.
      properties:
        bustime-response:
          type: object
          properties:
            directions:
              type: array
              description: >
                Array of direction objects. Note: in JSON the key is `directions` (plural),
                unlike the XML element name `dir`.
              items:
                $ref: '#/components/schemas/Direction'
            error:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'

    # ── STOP ──
    Stop:
      type: object
      description: A single bus stop.
      required:
        - stpid
        - stpnm
        - lat
        - lon
      properties:
        stpid:
          type: string
          description: Unique identifier for this stop.
          example: "1577"
        stpnm:
          type: string
          description: Display name of this stop (e.g. `"Madison and Clark"`).
          example: "1509 S Michigan"
        lat:
          type: number
          format: double
          description: Latitude position in decimal degrees (WGS 84).
          example: 41.861706666665
        lon:
          type: number
          format: double
          description: Longitude position in decimal degrees (WGS 84).
          example: -87.623969999999
        dtradd:
          type: array
          description: >
            List of detour IDs which temporarily add (service) this stop.
            Use `getdetours` to retrieve metadata about these detour IDs.
          items:
            type: string
          example: ["BFC46F62-990F-4AB4-A85C-3AF84574EC99"]
        dtrrem:
          type: array
          description: >
            List of detour IDs which remove (detour around) this stop.
            Use `getdetours` to retrieve metadata about these detour IDs.
          items:
            type: string
          example: ["50C633C7-0891-4E5A-83A8-FF0C6214BF69"]
        ada:
          type: boolean
          description: True if the stop is ADA accessible. Only included if supplied by the TA.
          example: true
    GetStopsResponse:
      type: object
      description: Root response wrapper for getstops.
      properties:
        bustime-response:
          type: object
          properties:
            stops:
              type: array
              description: >
                Array of stop objects. Note: in JSON the key is `stops` (plural),
                unlike the XML element name `stop`.
              items:
                $ref: '#/components/schemas/Stop'
            error:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'

    # ── PATTERN POINT ──
...
    Prediction:
      type: object
      description: A predicted arrival or departure time for a stop or vehicle.
      required:
        - tmstmp
        - typ
        - stpid
        - stpnm
        - vid
        - dstp
        - rt
        - rtdd
        - rtdir
        - des
        - prdtm
        - dyn
        - tablockid
        - tatripid
        - origtatripno
        - zone
        - psgld
        - flagstop
      properties:
        tmstmp:
          type: string
          description: >
            Date and time (local) when this prediction was generated.
            Format depends on `tmres`; Unix epoch ms if `unixTime=true`.
          example: "20250421 16:04"
        typ:
          type: string
          description: >
            Type of prediction.
            - `"A"` = Arrival (when vehicle will arrive at this stop)
            - `"D"` = Departure (when vehicle will depart — used for first stops and layovers)
          enum: [A, D]
          example: "A"
        stpid:
          type: string
          description: Unique identifier of the stop for which this prediction was generated.
          example: "456"
        stpnm:
          type: string
          description: Display name of the stop for which this prediction was generated.
          example: "Madison & Jefferson"
        vid:
          type: string
          description: Unique ID of the vehicle for which this prediction was generated.
          example: "8184"
        dstp:
          type: integer
          description: Linear distance in feet remaining for the vehicle to reach this stop.
          example: 686
        rt:
          type: string
          description: Alphanumeric designator of the route (e.g. `"20"` or `"X20"`).
          example: "20"
        rtdd:
          type: string
          description: Language-specific route designator intended for display.
          example: "20"
        rtdir:
          type: string
          description: Direction of travel of the route for this prediction (e.g. `"Eastbound"`).
          example: "Westbound"
        des:
          type: string
          description: Final destination of the vehicle associated with this prediction.
          example: "Austin"
        prdtm:
          type: string
          description: >
            Predicted date and time (local) of the vehicle's arrival or departure.
            Format depends on `tmres`; Unix epoch ms if `unixTime=true`.
          example: "20250421 16:06"
        dly:
          type: boolean
          description: True if the vehicle is delayed. Always present in v3.
          example: false
        dyn:
          type: integer
          description: >
            Dynamic action type affecting this prediction. Possible values:
            - `0`  = None (no change)
            - `1`  = Canceled
            - `2`  = Reassigned (moved to different vehicle/operator)
            - `3`  = Shifted (time moved)
            - `4`  = Expressed (drop-off only, will not pick up passengers)
            - `6`  = Stops Affected (some stops affected, but not the trip itself)
            - `8`  = New Trip (dynamically created, not in TA schedule)
            - `9`  = Partial Trip (trip split or short-turned)
            - `10` = Partial Trip New (split with new trip identifier)
            - `12` = Delayed Cancel (canceled but not shown to the public)
            - `13` = Added Stop (stop added, not originally scheduled)
            - `14` = Unknown Delay
            - `15` = Unknown Delay New (dynamic trip with delay)
            - `16` = Invalidated Trip (do not show predictions to public)
            - `17` = Invalidated Trip New (dynamic trip, invalidated)
            - `18` = Cancelled Trip New (dynamic trip, canceled)
            - `19` = Stops Affected New (dynamic trip, events affected)
          enum: [0, 1, 2, 3, 4, 6, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 19]
          example: 0
        tablockid:
          type: string
          description: TA's version of the scheduled block identifier for the current work being performed.
          example: "20 -803"
        tatripid:
          type: string
          description: TA's version of the scheduled trip identifier for the vehicle's current trip.
          example: "1040713"
        origtatripno:
          type: string
          description: Trip ID as defined by the TA scheduling system.
          example: "262522629"
        prdctdn:
          type: string
          description: >
            Time remaining in minutes until the bus arrives at this stop.
            Returns `"DUE"` when the bus is imminent.
          example: "DUE"
        zone:
          type: string
          description: Zone name if the vehicle has entered a defined zone, otherwise blank.
          example: ""
        psgld:
          type: string
          description: >
            Passenger load ratio. Values: `"FULL"`, `"HALF_EMPTY"`, `"EMPTY"`, `"N/A"`.
          enum: [FULL, HALF_EMPTY, EMPTY, "N/A", ""]
          example: ""
        stst:
          type: integer
          description: Scheduled trip start time in seconds past midnight.
          example: 57120
        stsd:
          type: string
          description: Scheduled trip start date in `yyyy-mm-dd` format.
          example: "2025-04-21"
        flagstop:
          type: integer
          description: >
            Flag stop information for this prediction.
            - `-1` = UNDEFINED (no flag-stop information available)
            - `0`  = NORMAL (normal stop)
            - `1`  = PICKUP_AND_DISCHARGE (flag stop for both pickup and discharge)
            - `2`  = ONLY_DISCHARGE (flag stop for discharge only)
          enum: [-1, 0, 1, 2]
          example: 0
        nbus:
          type: string
          description: Additional bus information, if applicable.
    GetPredictionsResponse:
      type: object
      description: Root response wrapper for getpredictions.
      properties:
        bustime-response:
          type: object
          properties:
            prd:
              type: array
              description: Array of prediction objects, sorted in ascending order by predicted time.
              items:
                $ref: '#/components/schemas/Prediction'
            error:
              type: array
              items:
                $ref: '#/components/schemas/ApiError'

    # ── LOCALE ──
...